separated collapsible Description component to a different file
