+++ ui/src/components/form/InputFileInput.jsx
@@ -1,8 +1,7 @@
 /* eslint-disable prettier/prettier */
-import { useState, useEffect, useRef } from "react";
+import { useState } from "react";
 import YAMLTextArea from "./YAMLTextArea";
 import { InputsDescription } from "../StepDescription";
-import ReactMarkdown from "react-markdown";
 import "./InputFileInputs.css";
 import ScriptInput from "./ScriptInput";
 import _lang from "lodash/lang";
@@ -10,12 +9,11 @@ import Box from "@mui/material/Box";
 import Tabs from "@mui/material/Tabs";
 import Tab from "@mui/material/Tab";
 import Alert from "@mui/material/Alert";
-import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-import ExpandLessIcon from "@mui/icons-material/ExpandLess";
 import InputType from "./InputType";
 
 import { styled } from "@mui/material";
 import { ScriptInputExample } from "./ScriptInputExample";
+import InputDescription from "./InputDescription";
 
 /**
  * An input that we use to fill the input file's content.
@@ -162,7 +160,7 @@ const InputForm = ({ inputs, inputFileContent, setInputFileContent }) => {
                   </>
                 </div>
 
-                <DescriptionSection
+                <InputDescription
                   description={description}
                   inputId={inputId}
                 />
@@ -174,95 +172,3 @@ const InputForm = ({ inputs, inputFileContent, setInputFileContent }) => {
   );
 };
 
-const DescriptionSection = ({ description, inputId }) => {
-  const descriptionCollapseThreshold = 100;
-  const [expanded, setExpanded] = useState(false);
-  const [canCollapse, setCanCollapse] = useState(false);
-  const [expandedHeight, setExpandedHeight] = useState(0);
-  const contentRef = useRef(null);
-
-  useEffect(() => {
-    const evaluateOverflow = () => {
-      if (!contentRef.current) {
-        setCanCollapse(false);
-        return;
-      }
-      const hasOverflow =
-        contentRef.current.scrollHeight > descriptionCollapseThreshold + 1;
-
-      setCanCollapse(hasOverflow);
-      if (!hasOverflow) {
-        setExpanded(false);
-      }
-    };
-
-    const animationFrameId = requestAnimationFrame(evaluateOverflow);
-    window.addEventListener("resize", evaluateOverflow);
-
-    return () => {
-      cancelAnimationFrame(animationFrameId);
-      window.removeEventListener("resize", evaluateOverflow);
-    };
-  }, [description, descriptionCollapseThreshold]);
-
-  const expandCollapse = () => {
-    setExpandedHeight(contentRef.current.scrollHeight);
-    setExpanded((oldValue) => !oldValue);
-  };
-
-  const isExpanded = expanded || !canCollapse;
-
-  return (
-    <div className="inputDescriptionWrapper">
-      <div
-        ref={contentRef}
-        className={`inputDescriptionContent ${isExpanded ? "expanded" : "collapsed"}`}
-        style={{
-          "--description-collapse-threshold": `${descriptionCollapseThreshold}px`,
-        }}
-      >
-        {description ? (
-          <div className="reactMarkdown"><ReactMarkdown>{description}</ReactMarkdown></div>
-        ) : (
-          <Alert severity="warning">
-            Missing description for input "{inputId}"
-          </Alert>
-        )}
-      </div>
-
-      {canCollapse && (
-        <div className="descriptionToggle">
-          <button
-            type="button"
-            className="descriptionToggleButton"
-            onClick={() => expandCollapse()}
-          >
-            {isExpanded ? (
-              <>
-                Show less{" "}
-                <ExpandLessIcon
-                  sx={{
-                    fontSize: "1.2rem",
-                    color: "var(--biab-green-main)",
-                    verticalAlign: "middle",
-                  }}
-                />
-              </>
-            ) : (
-              <>
-                Read more{" "}
-                <ExpandMoreIcon
-                  sx={{
-                    fontSize: "1.2rem",
-                    color: "var(--biab-green-main)",
-                    verticalAlign: "middle",
-                  }}
-                />
-              </>
-            )}
-          </button>
-        </div>
-      )}
-    </div>
-  );
-};
